Added 'SingleChildScrollView'
After adding transactions when it goes out of screen, we need something to make screen scrollable
Hence this widget was used to push the chart and the input fileds up when keyboard was raised
(especially when rotated).
Also, the transaction list was wrapped in container and then the widget was added here too.
By this, the scrolling of list was restricted to the height.